+++ gui/symfony-docker/src/Controller/ContactController.php
@@ -21,28 +21,6 @@ class ContactController extends AbstractController
         ]);
     }
 
-    #[Route('/report', name: 'app_report')]
-    public function report(Request $request, EntityManagerInterface $entityManager): Response
-    {
-        if (!$this->getUser()) {
-            return $this->redirectToRoute('app_login');
-        }
-        $report = new Report();
-        $report->setDate(new \DateTime('now', new \DateTimeZone('Europe/Paris')));
-        $report->setUser($this->getUser());
-        $report->setRead(false);
-        $form = $this->createForm(ReportType::class, $report);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->persist($report);
-            $entityManager->flush();
-            return $this->redirectToRoute('app_index');
-        }
-
-        return $this->render('contact/report.html.twig', [
-            'form' => $form->createView(),
-        ]);
-    }
