+++ gui/symfony-docker/templates/fragments/footer.html.twig
@@ -4,7 +4,7 @@
 
     <div class="row pt-4">
 
-      <a href="{{ path('app_about')}}" class="text-uppercase mb-4 font-weight-bold text-warning" style="font-size: 24px;">BC24</a>
+      <a href="{{ path('app_static_about')}}" class="text-uppercase mb-4 font-weight-bold text-warning" style="font-size: 24px;">BC24</a>
 
         <p><i>Chez BC24, nous révolutionnons la traçabilité de la viande en assurant une transparence totale depuis la ferme jusqu'à votre table, pour que vous puissiez savourer chaque bouchée en toute confiance.</i></p>
