+++ gui/symfony-docker/templates/fragments/header.html.twig
@@ -29,7 +29,7 @@
 						</a>
 					</li>
 					<li class="nav-item">
-						<a class="nav-link" href="{{ path('app_report') }}">Signaler un produit
+						<a class="nav-link" href="{{ path('app_report_reportAliment') }}">Signaler un produit
 
 						</a>
 					</li>
@@ -58,7 +58,7 @@
 
 					{% if app.user %}
 						<li class="nav-item">
-							<a class="btn-transparent" href="{{ path('app_myaccount') }}">Mon Compte</a>
+							<a class="btn-transparent" href="{{ path('app_user_account') }}">Mon Compte</a>
 						</li>
 						<li class="nav-item">
 							<a class="btn-transparent" href="{{ path('app_logout') }}">Déconnexion</a>
